I mistakenly added y.tab.c, will remove.

This set of changes fixes the 64-bit constant problems.

+++ util/6a/lex.c
@@ -1066,7 +1066,7 @@ zname(char *n, int t, int s)
 void
 zaddr(Gen *a, int s)
 {
-	long l;
+	int32 l;
 	int i, t;
 	char *n;
 	Ieee e;
